stripped old content out of load-rootfs.sh

+++ load-rootfs.sh
@@ -1,27 +1,20 @@
 #!/bin/bash
 
-#MNTROOTFS=/media/rootfs
 ORIG_PWD=${PWD}
 TARDIR=${PWD}/out/target/product/overo/
 
-#echo "Clean old rootfs on SD"
-#sudo rm -rf ${MNTROOTFS}/*
 cd ${TARDIR}
-echo "remove dir"
-rm -rf android_rootfs
-echo "make dir"
+if [ -e android_rootfs ]; then
+    echo "[remove old root directory]"
+    rm -rf android_rootfs
+fi
+echo "[make fresh root directory]"
 mkdir android_rootfs
-echo "move root"
+echo "[copy root file system]"
 cp -r root/* android_rootfs
-echo "move system"
+echo "[copy system folder]"
 cp -r system android_rootfs
-cd ${TARDIR}
-echo "make tarball"
+echo "[build tarball]"
 ../../../../build/tools/mktarball.sh ../../../host/linux-x86/bin/fs_get_stats android_rootfs . rootfs rootfs.tar.bz2
-#echo "untar"
-#sudo tar xaf rootfs.tar.bz2 --numeric-owner -C ${MNTROOTFS}
-#echo "sync"
-#sync
-#sync
 cd ${ORIG_PWD}
-echo "DONE!"
+echo "[DONE!]"
